requirements.txt
torch
transformers
Pillow
accelerate
sentencepiece
protobuf
einops
pyvips